requirements.txt
RPi.GPIO>=0.7.1
Flask>=2.3.0
Flask-SocketIO>=5.3.0
paramiko>=3.0.0
python-socketio>=5.8.0
eventlet>=0.33.0
bcrypt>=4.0.0
